SF Bug #2053299: ProjectViewer 2.9: Confused about symlinks
It's not really PV who is confused, but combination of VFS reading the link before the real path, and the symlink loop detection code being way too strict. The loop detection logic should only be applied when resolving symlinks; symlinks that point to other directories in the same tree but don't form a loop can be safely traversed.

This causes the traversal to return more data than before in certain cases, but it's returning the correct data given the user's filesystem.